requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
dbus-python==1.2.12
Django==3.1
djangorestframework==3.11.1
docopt==0.6.2
idna==2.8
itsdangerous==1.1.0
lxml==4.4.2
MarkupSafe==1.1.1
Pillow==7.2.0
psutil==5.6.7
pycairo==1.18.2
PyGObject==3.34.0
python-xlib==0.26
pytz==2020.1
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.1
ueberzug==18.1.5
Unidecode==1.1.1
urllib3==1.25.7
yet-another-spotify-lyrics==2.3.1
